Leaf 173: modularize OK body controls and lock newline spacing
